Fix tests on Python 2
ZODB 5.5.0 added an "import mock" to ZODB.tests.util (or "from unittest
import mock" on Python 3, which is in the standard library).  ZODB[test]
included the new dependency, but zope.component wasn't using that for
tests.

However, ZODB.tests.util.DB was just an import of
ZODB.MappingStorage.DB, so we can just as easily use the latter and
avoid the extra dependency.